refactored flow layout and added memory management
