+++ GHFollowers/UIHelpers/CollectionViewFlowLayoutHelper.swift
@@ -1,0 +1,24 @@
+//
+//  CollectionViewFlowLayoutHelper.swift
+//  GHFollowers
+//
+//  Created by Nathan Tugwell on 13/07/2022.
+//
+
+import UIKit
+
+struct CollectionViewLayoutHelper {
+
+    static func createFlowLayout(in view: UIView) -> UICollectionViewFlowLayout {
+        let width = view.bounds.width
+        let padding: CGFloat = 12
+        let minimumItemSpacing: CGFloat = 10
+        let availableWidth = width - (padding * 2) - (minimumItemSpacing * 2)
+        let itemWidth = availableWidth / 3
+        
+        let flowLayout = UICollectionViewFlowLayout()
+        flowLayout.sectionInset = UIEdgeInsets(top: padding, left: padding, bottom: padding, right: padding)
+        flowLayout.itemSize = CGSize(width: itemWidth, height: itemWidth + 40)
+        return flowLayout
+    }
+}
